Refactor sourceWritingStopAlgorithm and checkoutLockAlgorithm SPI as JobLock algorithm
