+++ shardingsphere-kernel/shardingsphere-data-pipeline/shardingsphere-data-pipeline-core/src/main/resources/META-INF/services/org.apache.shardingsphere.data.pipeline.spi.lock.RowBasedJobLockAlgorithm
@@ -15,4 +15,4 @@
 # limitations under the License.
 #
 
-org.apache.shardingsphere.data.pipeline.scenario.rulealtered.spi.DefaultRuleAlteredSourceWritingStopAlgorithm
+org.apache.shardingsphere.data.pipeline.scenario.rulealtered.spi.DefaultSourceWritingStopAlgorithm
